DUOS-925 - Remove Election APIs in lieu of ElectionService
