+++ src/test/java/org/broadinstitute/consent/http/resources/ElectionResourceTest.java
@@ -4,8 +4,7 @@ import com.google.api.client.http.HttpStatusCodes;
 import org.apache.commons.lang3.RandomStringUtils;
 import org.apache.commons.lang3.RandomUtils;
 import org.broadinstitute.consent.http.models.Election;
-import org.broadinstitute.consent.http.service.AbstractElectionAPI;
-import org.broadinstitute.consent.http.service.ElectionAPI;
+import org.broadinstitute.consent.http.service.ElectionService;
 import org.broadinstitute.consent.http.service.VoteService;
 import org.junit.Assert;
 import org.junit.Before;
@@ -32,7 +31,6 @@ import static org.mockito.Mockito.when;
 @SuppressWarnings("FieldCanBeLocal")
 @RunWith(PowerMockRunner.class)
 @PowerMockIgnore("jdk.internal.reflect.*")
-@PrepareForTest({AbstractElectionAPI.class})
 public class ElectionResourceTest {
 
     private final int OK = HttpStatusCodes.STATUS_CODE_OK;
@@ -43,25 +41,23 @@ public class ElectionResourceTest {
     VoteService voteService;
 
     @Mock
-    ElectionAPI electionAPI;
+    ElectionService electionService;
 
     private ElectionResource electionResource;
 
     @Before
     public void setUp() {
         MockitoAnnotations.initMocks(this);
-        PowerMockito.mockStatic(AbstractElectionAPI.class);
         when(voteService.findVotesByReferenceId(any())).thenReturn(Collections.emptyList());
         doNothing().when(voteService).advanceVotes(any(), anyBoolean(), anyString());
-        when(electionAPI.checkDataOwnerToCloseElection(any())).thenReturn(false);
-        doNothing().when(electionAPI).closeDataOwnerApprovalElection(any());
-        when(electionAPI.updateElectionById(any(), any())).thenReturn(new Election());
-        when(electionAPI.describeElectionById(any())).thenReturn(new Election());
-        when(electionAPI.describeElectionByVoteId(any())).thenReturn(new Election());
-        when(electionAPI.isDataSetElectionOpen()).thenReturn(true);
-        when(electionAPI.getConsentElectionByDARElectionId(any())).thenReturn(new Election());
-        when(AbstractElectionAPI.getInstance()).thenReturn(electionAPI);
-        electionResource = new ElectionResource(voteService);
+        when(electionService.checkDataOwnerToCloseElection(any())).thenReturn(false);
+        doNothing().when(electionService).closeDataOwnerApprovalElection(any());
+        when(electionService.updateElectionById(any(), any())).thenReturn(new Election());
+        when(electionService.describeElectionById(any())).thenReturn(new Election());
+        when(electionService.describeElectionByVoteId(any())).thenReturn(new Election());
+        when(electionService.isDataSetElectionOpen()).thenReturn(true);
+        when(electionService.getConsentElectionByDARElectionId(any())).thenReturn(new Election());
+        electionResource = new ElectionResource(voteService, electionService);
     }
 
     @Test
@@ -74,7 +70,7 @@ public class ElectionResourceTest {
     @Test
     public void testAdvanceElectionError() {
         when(voteService.findVotesByReferenceId(anyString())).thenThrow(new NotFoundException());
-        electionResource = new ElectionResource(voteService);
+        electionResource = new ElectionResource(voteService, electionService);
         String referenceId = RandomStringUtils.random(10);
         Response response = electionResource.advanceElection(referenceId, "Yes");
         Assert.assertEquals(NOT_FOUND, response.getStatus());
@@ -88,8 +84,8 @@ public class ElectionResourceTest {
 
     @Test
     public void testUpdateElectionError() {
-        when(electionAPI.updateElectionById(any(), anyInt())).thenThrow(new NotFoundException());
-        electionResource = new ElectionResource(voteService);
+        when(electionService.updateElectionById(any(), anyInt())).thenThrow(new NotFoundException());
+        electionResource = new ElectionResource(voteService, electionService);
         Response response = electionResource.updateElection(new Election(), randomInt());
         Assert.assertEquals(NOT_FOUND, response.getStatus());
     }
@@ -102,8 +98,8 @@ public class ElectionResourceTest {
 
     @Test
     public void testDescribeElectionByIdError() {
-        when(electionAPI.describeElectionById(anyInt())).thenThrow(new NotFoundException());
-        electionResource = new ElectionResource(voteService);
+        when(electionService.describeElectionById(anyInt())).thenThrow(new NotFoundException());
+        electionResource = new ElectionResource(voteService, electionService);
         Response response = electionResource.describeElectionById(randomInt());
         Assert.assertEquals(NOT_FOUND, response.getStatus());
     }
@@ -116,8 +112,8 @@ public class ElectionResourceTest {
 
     @Test
     public void testDescribeElectionByVoteIdError() {
-        when(electionAPI.describeElectionByVoteId(anyInt())).thenThrow(new NotFoundException());
-        electionResource = new ElectionResource(voteService);
+        when(electionService.describeElectionByVoteId(anyInt())).thenThrow(new NotFoundException());
+        electionResource = new ElectionResource(voteService, electionService);
         Response response = electionResource.describeElectionByVoteId(randomInt());
         Assert.assertEquals(NOT_FOUND, response.getStatus());
     }
@@ -130,8 +126,8 @@ public class ElectionResourceTest {
 
     @Test
     public void testIsDataSetElectionOpenError() {
-        when(electionAPI.isDataSetElectionOpen()).thenThrow(new NullPointerException());
-        electionResource = new ElectionResource(voteService);
+        when(electionService.isDataSetElectionOpen()).thenThrow(new NullPointerException());
+        electionResource = new ElectionResource(voteService, electionService);
         Response response = electionResource.isDataSetElectionOpen(null);
         Assert.assertEquals(ERROR, response.getStatus());
     }
